+++ src/Filament/V5/Resources/PostResource/Pages/ListPosts.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kurt\Modules\Blog\Filament\V5\Resources\PostResource\Pages;
+
+use Filament\Actions\Action;
+use Filament\Actions\CreateAction;
+use Filament\Resources\Pages\ListRecords;
+use Kurt\Modules\Blog\Filament\V5\Resources\PostResource;
+
+class ListPosts extends ListRecords
+{
+    protected static string $resource = PostResource::class;
+
+    /**
+     * @return array<int, Action>
+     */
+    protected function getHeaderActions(): array
+    {
+        return [
+            CreateAction::make(),
+        ];
+    }
+}